+++ src/Symfony/Component/Notifier/Bridge/Novu/Tests/NovuTransportTest.php
@@ -37,7 +37,7 @@ class NovuTransportTest extends TransportTestCase
 
     public static function supportedMessagesProvider(): iterable
     {
-        yield [new PushMessage('test', '{}', new NovuOptions(123, null, null, 'test@example.com', null, null, null, ['email' => ['from' => 'no-reply@example.com', 'senderName' => 'No-Reply']], []))];
+        yield [new PushMessage('test', '{}', new NovuOptions(123, null, null, 'test@example.com', null, null, null, ['email' => ['from' => 'no-reply@example.com', 'senderName' => 'No-Reply']], [], ['tenant' => 'tenant-id', 'app' => 'app-id']))];
     }
 
     public static function unsupportedMessagesProvider(): iterable
@@ -55,6 +55,6 @@ class NovuTransportTest extends TransportTestCase
         $this->expectException(TransportException::class);
         $this->expectExceptionMessageMatches('/400: "subscriberId under property to is not configured"/');
 
-        $transport->send(new PushMessage('test', '{}', new NovuOptions(123, null, null, 'test@example.com', null, null, null, ['email' => ['from' => 'no-reply@example.com', 'senderName' => 'No-Reply']], [])));
+        $transport->send(new PushMessage('test', '{}', new NovuOptions(123, null, null, 'test@example.com', null, null, null, ['email' => ['from' => 'no-reply@example.com', 'senderName' => 'No-Reply']], [], ['tenant' => 'tenant-id', 'app' => 'app-id'])));
     }
 }