+++ src/Symfony/Component/Notifier/Bridge/Novu/CHANGELOG.md
@@ -1,6 +1,12 @@
 CHANGELOG
 =========
 
+8.2
+---
+
+ * Add `$context` parameter to `NovuOptions`
+ * Deprecate `NovuSubscriberRecipient::getOverrides()` and its `$overrides` constructor parameter, pass overrides to `NovuOptions` instead
+
 6.4
 ---